+++ src/Components/MainContent/Components/Activities/activites.css
