+++ src/Components/MainContent/Components/Competition/competition.css
