+++ src/Components/MainContent/Components/Competition/competition.tsx
