+++ src/Components/MainContent/Components/Scholarship/scholarship.tsx
@@ -1,0 +1,16 @@
+// 장학 , 폐기 예정 
+import React from 'react'
+import './scholarshop.css'
+
+const Scholarship: React.FC = () => {
+
+  return (
+    <div className="scholar-all-container">
+      <div className="later-card">
+        <img src="img/404.svg" className="errorimg" />
+      </div>
+    </div>
+  )
+}
+
+export default Scholarship
