+++ src/Components/MainContent/Container/Contentselector.tsx
@@ -1,17 +1,17 @@
 import React, { useState, useEffect } from 'react'
 import './Content.css'
-import Noticelist from './Notice/noticelist'
-import Noticemain from './Notice/noticemain'
-import Activities from './Activities/activities'
-import SearchBox from '../Search/Searchbox'
-import Scholarship from './Scholarship/scholarship'
-import Competition from './Competition/competition'
-import Department from './Department/Department'
-import RecruitmentContainer from './Recruitment/recruitment'
-import Job from './job/job'
-import { UserData } from '../../api'
-import Mypageinfo from '../Modal/MyPageInfoModal/Mypageinfomodal'
-import { getStorageData, setStorageData } from '../../util/storage'
+import Noticelist from '../Components/Notice/noticelist'
+import Noticemain from '../Components/Notice/noticemain'
+import Activities from '../Components/Activities/activities'
+import SearchBox from '../../Search/Searchbox'
+import Scholarship from '../Components/Scholarship/scholarship'
+import Competition from '../Components/Competition/competition'
+import Department from '../Components/Department/Department'
+import RecruitmentContainer from '../Components/Recruitment/recruitment'
+import Job from '../Components/job/job'
+import { UserData } from '../../../api'
+import Mypageinfo from '../../Modal/MyPageInfoModal/Mypageinfomodal'
+import { getStorageData, setStorageData } from '../../../util/storage'
 interface ContentSelectorProps {
   userData: UserData | null
 }
